{{cookiecutter.project_name}}/requirements.txt
PyYAML==4.2b1
colorama
